+++ app/controllers/healthcheck_controller.rb
@@ -1,0 +1,7 @@
+class HealthcheckController < ApplicationController
+  skip_before_action :redirect_unauthenticated_user
+
+  def check
+    render json: {status: "OK"}, status: :ok
+  end
+end