+++ config/routes.rb
@@ -103,6 +103,8 @@ Rails.application.routes.draw do
   get "cookies", to: "cookies#edit"
   post "cookies", to: "cookies#update"
 
+  get "healthcheck" => "healthcheck#check"
+
   # High voltage configuration for static pages. Matches routes from the root of the domain. Uses
   # HighVoltage::Constraints::RootRoute to validate that the view exists.
   get "/*id" => "pages#show", :as => :page, :format => false, :constraints => HighVoltage::Constraints::RootRoute.new